debugging

+++ bin/apply-multi-configs.sh
@@ -149,9 +149,11 @@ rm -f  "$waitFile" "$onlyCasesFile"
 cat "$configsListFile" | while read prefix; do
     confId=$(basename "$prefix")
     id="${outputIdPrefix}${confId}"
-    cat "$destDir/$id/predicted.answers" >"$destDir/$id.answers"
-    echo "$progName DEBUG: predicted answers in '$destDir/$id.answers', removing '$destDir/$id'" 1>&2
-    rm -rf "$destDir/$id" # update oct 16: remove dir to save space
+    if [ $resume -eq 0 ] || [ ! -s "$destDir/$id.answers" ]; then
+	cat "$destDir/$id/predicted.answers" >"$destDir/$id.answers"
+	echo "$progName DEBUG: predicted answers in '$destDir/$id.answers', removing '$destDir/$id'" 1>&2
+	rm -rf "$destDir/$id" # update oct 16: remove dir to save space
+    fi
 done
 
 echo "$progName: done."
